requirements.txt
# core
python-dotenv>=1.0.0
typer[all]>=0.9.0
requests>=2.28.0
tqdm>=4.65.0
pandas>=1.5.0
numpy>=1.24.0

# transformers & tokenizers
transformers>=4.30.0
tokenizers>=0.13.3
datasets>=2.12.0

# torch
torch>=2.0.0

# for simple GNN-like ops (we implement pure torch message passing)
# optionally, if you want better GNNs:
# torch-geometric (manual install recommended)
# dgl, pyg: optional

# code parsers
tree_sitter>=0.20.1
gitpython>=3.1.31

# dev / utils
scikit-learn>=1.2.0
black
